refactor(dashboard): extract useDashboardModules hook + shared TileFace
Pull the data/role orchestration out of Dashboard.tsx into a presentation-free
useDashboardModules() hook, and split the bento tile's visual contents into a
shared TileFace. No behavior change — Dashboard still renders the in-place bento
overview — but both the hook and the face are now reusable by a route-based
overview. Sets up the /dashboard/[module] structure without duplicating the
gating logic or the tile markup.
